backend/requirements.txt
# Backend Python dependencies

# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.0

# Database & ORM
sqlmodel==0.0.22
psycopg2-binary==2.9.10
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Validation & serialization
pydantic==2.9.0
pydantic-settings==2.5.0
email-validator==2.1.1

# Environment
python-dotenv==1.0.1

# CORS
python-multipart==0.0.12

# Testing
pytest==8.3.0
pytest-asyncio==0.24.0
httpx==0.27.0

# AI Integration
groq==1.0.0
